Confirmation mail

+++ src/components/Organization/organization-tabpane.tsx
@@ -5,6 +5,7 @@ import { Administrator, Company } from 'src/api/types';
 import { Util } from 'src/helpers/util';
 import { useAppDispatch } from 'src/redux/hooks';
 import { refreshCompanies } from 'src/redux/slices/companies/companies.slice';
+import { confirmation } from '../Modals/ConfirmationModal.component';
 import { OrganizationTable } from '../Table/organization-table';
 
 interface IOrganizationTabPane {
@@ -14,6 +15,7 @@ interface IOrganizationTabPane {
 export const OrganizationTabPane = (props: IOrganizationTabPane) => {
   const { trigger } = props;
   const [loading, setLoading] = useState(false);
+
   const [{ data, meta }, setData] = useState({
     data: [] as Administrator[],
     meta: Util.getDefaultPaginationMeta({}),
@@ -35,25 +37,30 @@ export const OrganizationTabPane = (props: IOrganizationTabPane) => {
   }, [setLoading, setData, query]);
 
   const deleteOrganization = async (id: string) => {
-    const toast = (await import('react-toastify')).toast;
     if (!loading) {
-      setLoading(true);
-      const clone = data.map((d) => ({ ...d }));
-      try {
-        setData({
-          meta,
-          data: data.filter((d) => (d.company as Company).id !== id),
-        });
-        await $api.company.deleteCompany(id);
-        getOrganizations();
-        refreshCompanies(dispatch);
-        toast.success('company deleted successfully');
-      } catch (error) {
-        const err = error as HttpError;
-        toast.error(err.message);
-        setData({ meta, data: clone });
-      } finally {
-        setLoading(false);
+      const shouldDelete = await confirmation({
+        text: 'Are you sure you want to permanently delete this organisation?',
+      });
+      if (shouldDelete) {
+        const toast = (await import('react-toastify')).toast;
+        setLoading(true);
+        const clone = data.map((d) => ({ ...d }));
+        try {
+          setData({
+            meta,
+            data: data.filter((d) => (d.company as Company).id !== id),
+          });
+          await $api.company.deleteCompany(id);
+          getOrganizations();
+          refreshCompanies(dispatch);
+          toast.success('company deleted successfully');
+        } catch (error) {
+          const err = error as HttpError;
+          toast.error(err.message);
+          setData({ meta, data: clone });
+        } finally {
+          setLoading(false);
+        }
       }
     }
   };
